+++ global_trotter_web/src/services/api.js
@@ -1,7 +1,7 @@
 import { getToken } from './tokenStorage.js'
 
 
-export const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || 'http://10.42.60.127:5000'//'http://192.168.1.100:5000'
+export const API_BASE_URL = import.meta.env.VITE_API_BASE_URL || 'http://192.168.1.100:5000'
 
 export async function apiRequest(endpoint, options = {}) {
   const token = getToken()